+++ .dev-tools/utils/createJestConfig.js
@@ -24,8 +24,6 @@ module.exports = (resolve, rootDir) => {
       '^activities(.*)$': '<rootDir>/activities$1',
       '^client(.*)$': '<rootDir>/client$1',
       '^core(.*)$': '<rootDir>/core$1',
-      // Maps
-      '.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/.dev-tools/config/jest/fileTransform.js',
       // FIXME @see https://github.com/facebook/jest/issues/553
       'vendors/blockly/blocks_compressed': `${paths.devElements}/test/__mocks__/blockly/blocks_compressed.js`,
       'vendors/blockly/javascript_compressed': `${paths.devElements}/test/__mocks__/blockly/javascript_compressed.js`,
@@ -37,6 +35,11 @@ module.exports = (resolve, rootDir) => {
     ],
     testRegex: '(activities|client|core)/.*\\.test\\.jsx?$',
     testURL: 'http://localhost',
+    transform: {
+      '^.+\\.jsx?$': 'babel-jest',
+      // Image Files
+      '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/.dev-tools/config/jest/fileTransform.js',
+    },
   };
 
   if (rootDir) {