+++ client/components/pages/links/HeaderLink.jsx
@@ -4,6 +4,8 @@
  *
  * @flow
  */
+// @see https://github.com/styled-components/stylelint-processor-styled-components/issues/54
+// stylelint-disable
 import React from 'react';
 import { Link as RouterLink } from 'react-router-dom';
 import styled from 'styled-components';
@@ -13,6 +15,7 @@ import { COLOR_PALETTE } from 'core/constants/colors';
 const LinkContainer = styled.li`
   list-style-type: none;
   position: relative;
+
   &::before {
     content: '';
     height: 100%;
@@ -22,11 +25,13 @@ const LinkContainer = styled.li`
     width: 0;
     z-index: -1;
   }
+
   &:hover::before {
     background: ${COLOR_PALETTE.orange.clear};
     left: 0;
     width: 100%;
   }
+
   &::after {
     bottom: 0;
     content: '';
@@ -36,6 +41,7 @@ const LinkContainer = styled.li`
     transition: 0.5s;
     width: 0;
   }
+
   &:hover::after {
     background: ${COLOR_PALETTE.orange.light};
     left: 0;
@@ -51,6 +57,7 @@ const Link = styled(RouterLink)`
   height: 100%;
   padding: 0 0.5em;
   text-decoration: none;
+
   ${LinkContainer}:hover > a {
     font-weight: bold;
   }
@@ -71,6 +78,7 @@ const ChildrenContainer = styled.ul`
   visibility: hidden;
   width: 100%;
   z-index: -1;
+
   ${LinkContainer}:hover > & {
     opacity: 1;
     transform: translateY(0%);