+++ core/ui/BlocklyApp/components/Button.stories.jsx
@@ -15,6 +15,7 @@ const Container = styled.div`
   display: flex;
   flex-wrap: wrap;
   padding: 50px 0 0;
+
   & a {
     margin: 20px;
   }