+++ activities/NumericLine/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## HEAD
 
+- 🎨 Add an image for the `MazeWrongExitError`.
+- 🎨 Improve game messages.
+
 ## v1.0.1
 
 - Add SVG image instead of emoji to `leave_maze` block. This homogenize the icon through every OS and browser, and ensures is going to be rendered as expected.